Separate proxy and veRL checkpoint reload extensions
