+++ experimental/lite/examples/verl/verl_mlite/rollout/vllm_worker.py
@@ -1,16 +1,10 @@
 # Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-"""vLLM checkpoint-format weight reload extension for veRL."""
+"""Dependency-free vLLM checkpoint reload helpers and proxy extension."""
 
 from __future__ import annotations
 
-import os
 from collections.abc import Callable
 from typing import Any
-
-try:
-    from verl.workers.rollout.vllm_rollout.utils import vLLMColocateWorkerExtension
-except ModuleNotFoundError:
-    vLLMColocateWorkerExtension = object
 
 
 def _runner_model(model_runner: Any) -> Any:
@@ -48,42 +42,16 @@ def reload_checkpoint_buckets(
     finalize(model, _runner_model_config(model_runner))
 
 
-class VllmCheckpointWorkerExtension(vLLMColocateWorkerExtension):
-    """Consume already-serialized checkpoint tensors without online quantization."""
+class VllmCheckpointPathWorkerExtension:
+    """Minimal vLLM-only extension used by the single-GPU proxy."""
 
     def __new__(cls, **kwargs):
-        if os.environ.get("VERL_VLLM_FP8_QUANT_ENABLED", "0") == "1":
-            raise RuntimeError(
-                "vllm_checkpoint resync is incompatible with veRL online FP8 quantization; "
-                "set VERL_VLLM_FP8_QUANT_ENABLED=0"
-            )
-        return super().__new__(cls, **kwargs)
-
-    def update_weights_from_ipc(
-        self,
-        peft_config: dict | None = None,
-        base_sync_done: bool = False,
-        use_shm: bool = False,
-    ) -> None:
-        if peft_config or base_sync_done:
-            raise NotImplementedError(
-                "checkpoint-format resync does not support LoRA delta updates"
-            )
-
-        from verl.workers.rollout.vllm_rollout.bucketed_weight_transfer import (
-            BucketedWeightReceiver,
-        )
-
-        if self.device is None:
-            raise RuntimeError("vLLM worker device is not initialized")
-        receiver = BucketedWeightReceiver(
-            zmq_handle=self._get_zmq_handle(), device=self.device, use_shm=use_shm
-        )
-        reload_checkpoint_buckets(self.model_runner, receiver.receive_weights)
+        del kwargs
+        return super().__new__(cls)
 
     def reload_checkpoint_from_path(self, path: str) -> None:
         """Reload a serialized checkpoint directory for validation or recovery."""
         self.model_runner.reload_weights(weights_path=path, is_checkpoint_format=True)
 
 
-__all__ = ["VllmCheckpointWorkerExtension", "reload_checkpoint_buckets"]
+__all__ = ["VllmCheckpointPathWorkerExtension", "reload_checkpoint_buckets"]
